streamlit-langchain-tutorials/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv
faiss-cpu
pypdf
chromadb
tiktoken
